backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.2.3
scikit-learn==1.2.2
